feat: add TouchpanelBase class and initial implementation
